Remove deprecated Mail::CheckDeliveryParams for mail 2.9.0+ compatibility (#5)
* Remove deprecated Mail::CheckDeliveryParams dependency for mail 2.9.0+ compatibility

The Mail gem removed the CheckDeliveryParams module in version 2.9.0 after deprecation.
This module was included but never actually used by the DeliveryMethod class.

This change removes both the require and include statements to maintain compatibility
with mail 2.9.0 and later versions.

* Update .rubocop.yml to fix deprecated cop names

- Change 'require' to 'plugins' for rubocop extensions
- Update deprecated cop names:
  - Metrics/LineLength -> Layout/LineLength
  - Documentation -> Style/Documentation
  - Style/FileName -> Naming/FileName
  - Style/MethodLength -> Metrics/MethodLength
  - Style/AccessorMethodName -> Naming/AccessorMethodName

This fixes rubocop config errors and allows linting to run successfully.